Split `WithReadTimeout` into sub-crate.
