+++ src/lib.rs
@@ -61,11 +61,7 @@ extern crate time;
 extern crate num;
 #[macro_use] extern crate log;
 #[cfg(test)] extern crate quickcheck;
-
-#[cfg(unix)]
-extern crate nix;
-#[cfg(windows)]
-extern crate libc;
+extern crate with_read_timeout;
 
 // Public API
 pub use socket::UtpSocket;
